refactor(app): dedupe terminal url helpers and tests
Reuse resolveTerminalApiOriginUrl from terminal.ts for image fetch URL
resolution to eliminate the duplicate origin builder, and merge the
image fetch URL test cases into terminal.test.ts so they share the
existing mock setup. Extract a same-origin location stub helper to
keep individual test bodies short. Also extract attachGlobalResizeListeners
out of mountTerminalSession to stay under the max-lines-per-function
threshold.
